style: controllers 파일 병합
